+++ nodes/MortgageBE/IInputType.ts
@@ -1,0 +1,6 @@
+export default interface IInputType {
+  duration: string;
+  amount: string;
+  type: string;
+  locale: string;
+}
