+++ nodes/MortgageBE/IParsedInput.ts
@@ -1,0 +1,6 @@
+export default interface IParsedInput {
+    duration: string;
+    amount: number;
+    type: string;
+    country: string;
+}
