+++ nodes/MortgageBE/IProcessData.ts
@@ -1,0 +1,6 @@
+import IParsedInput from "./IParsedInput";
+
+export default interface IProcessedData {
+    originalInput: IParsedInput;
+    matchingRows: any[];
+}
